refactor: merged LineWidth to VectorLine
