+++ sites/geohub/src/components/maplibre/line/VectorLine.svelte
@@ -1,18 +1,25 @@
 <script lang="ts">
-	import LineWidth from '$components/maplibre/line/LineWidth.svelte';
+	import { page } from '$app/stores';
 	import VectorColorClassification from '$components/maplibre/vector/VectorColorClassification.svelte';
 	import VectorSimulationAccordion from '$components/maplibre/vector/VectorSimulationAccordion.svelte';
+	import VectorValueClassification from '$components/maplibre/vector/VectorValueClassification.svelte';
 	import type { Tag } from '$lib/types';
+	import { DEFAULTCOLOR_CONTEXT_KEY, type DefaultColorStore } from '$stores';
 	import {
 		Accordion,
 		Help,
 		LinePattern,
 		type VectorTileMetadata
 	} from '@undp-data/svelte-undp-components';
+	import { getContext } from 'svelte';
+
+	const defaultColorStore: DefaultColorStore = getContext(DEFAULTCOLOR_CONTEXT_KEY);
 
 	export let layerId: string;
 	export let metadata: VectorTileMetadata;
 	export let tags: Tag[];
+
+	let defaultLineWidth = $page.data.config.LineWidth;
 
 	let expanded: { [key: string]: boolean } = {
 		'line-color': true
@@ -53,7 +60,18 @@
 
 <Accordion title="Line width" bind:isExpanded={expanded['line-width']}>
 	<div class="pb-2" slot="content">
-		<LineWidth {layerId} {metadata} />
+		<VectorValueClassification
+			{layerId}
+			{metadata}
+			bind:defaultValue={defaultLineWidth}
+			minValue={0}
+			maxValue={10}
+			stepValue={0.1}
+			propertyName={'line-width'}
+			styleType="paint"
+			legendCssTemplate={`margin-top: 13px; width: 40px; height: {value}px; background-color: ${$defaultColorStore};`}
+			dataLabel="Line width"
+		/>
 	</div>
 	<div slot="buttons">
 		<Help>Stroke thickness.</Help>
